requirements.txt
requests==2.32.*
python-dotenv==1.0.*
argparse==1.4.*
python-telegram-bot==13.15
urllib3==1.26.6